chore: move common component to components
