+++ packages/suite/src/components/wallet/EstimatedMiningTime/index.tsx
